requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
appdirs==1.4.4
    # via requests-cache
asgiref==3.5.0
    # via django
async-timeout==4.0.2
    # via redis
attrs==21.4.0
    # via
    #   cattrs
    #   requests-cache
cattrs==1.10.0
    # via requests-cache
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.12
    # via requests
deprecated==1.2.13
    # via redis
django==4.0.4
    # via -r requirements.in
idna==3.3
    # via requests
packaging==21.3
    # via redis
psycopg2==2.9.3
    # via -r requirements.in
pyparsing==3.0.8
    # via packaging
python-dotenv==0.20.0
    # via -r requirements.in
redis==4.2.2
    # via -r requirements.in
requests==2.27.1
    # via
    #   -r requirements.in
    #   requests-cache
requests-cache==0.9.4
    # via -r requirements.in
six==1.16.0
    # via url-normalize
sqlparse==0.4.2
    # via django
url-normalize==1.4.3
    # via requests-cache
urllib3==1.26.9
    # via
    #   requests
    #   requests-cache
wrapt==1.14.1
    # via deprecated
